Final Marques Tree (unpolished)
look around, look around at how lucky we are to be alive right now - look around, look around, at how lucky we are to be alive right now.

history is happening in manHATTAN and we just HAPPEN to be in the greatest city in the world, in the greatest city in world!